delete converters for Ref and Struct, logging for exceptions in coverter methods
